+++ src/services/map/index.js
@@ -1,7 +1,7 @@
 const Json = require('../json');
 const makeObject = require('../../handlers/make/makeObject');
 
-const translateService = require('./translate');
+const transformService = require('./transform');
 
 const loadDefaultSettings = require('../../settings/loadDefaultSettings');
 
@@ -22,6 +22,7 @@ class Map {
   }
 
   /**
+   * Deprecated:
    * Translate a single value into destination object. Query will stop after first match.
    * @param {String} originPath - path from where data should be obtained from origin object
    * @param {originPath} destinationPath - path to where data should be mapped into destination
@@ -29,10 +30,13 @@ class Map {
    * @param {Object} functions - object of functions that can be called within query.
    */
   translate(originPath, destinationPath, functions) {
-    return translateService('translateSingle', originPath, destinationPath, functions, this);
+    console.warn('Deprecation warning: The use of translate is deprecated and will be removed from version 1.0.0 onwards. Use transform with { resolveOne: true, buildOne: true } settings instead.');
+    return transformService(originPath, destinationPath, functions,
+      { ...this, settings: { ...this.settings, resolveOne: true, buildOne: true } });
   }
 
   /**
+   * Deprecated:
    * Translate all values into destination object. Destination will be an array with all results
    * from origin query.
    * @param {String} originPath - path from where data should be obtained from origin object
@@ -41,10 +45,12 @@ class Map {
    * @param {Object} functions - object of functions that can be called within query.
    */
   translateAll(originPath, destinationPath, functions) {
-    return translateService('translateAll', originPath, destinationPath, functions, this);
+    console.warn('Deprecation warning: The use of translateAll is deprecated and will be removed from version 1.0.0 onwards. Use transform instead.');
+    return transformService(originPath, destinationPath, functions, this);
   }
 
   /**
+   * Deprecated:
    * Translate single value into destination object. Destination will be the first matching value
    * from origin path, mapped into all qualified destinations.
    * @param {String} originPath - path from where data should be obtained from origin object
@@ -53,10 +59,13 @@ class Map {
    * @param {Object} functions - object of functions that can be called within query.
    */
   translateOneToAll(originPath, destinationPath, functions) {
-    return translateService('translateOneToAll', originPath, destinationPath, functions, this);
+    console.warn('Deprecation warning: The use of translateOneToAll is deprecated and will be removed from version 1.0.0 onwards. Use transform with { resolveOne: true } setting instead.');
+    return transformService(originPath, destinationPath, functions,
+      { ...this, settings: { ...this.settings, resolveOne: true } });
   }
 
   /**
+   * Deprecated:
    * Translate all values into destination object. The single destination will be filled with an
    * array filled with all results from the originPath query.
    * @param {String} originPath - path from where data should be obtained from origin object
@@ -65,7 +74,20 @@ class Map {
    * @param {Object} functions - object of functions that can be called within query.
    */
   translateAllToOne(originPath, destinationPath, functions) {
-    return translateService('translateAllToOne', originPath, destinationPath, functions, this);
+    console.warn('Deprecation warning: The use of translateAllToOne is deprecated and will be removed from version 1.0.0 onwards. Use transform with { buildOne: true } setting instead.');
+    return transformService(originPath, destinationPath, functions,
+      { ...this, settings: { ...this.settings, buildOne: true } });
+  }
+
+  /**
+   * Transform values from origin object into destination object.
+   * @param {String} originPath - path from where data should be obtained from origin object
+   * @param {originPath} destinationPath - path to where data should be mapped into destination
+   * object
+   * @param {Object} functions - object of functions that can be called within query.
+   */
+  transform(originPath, destinationPath, functions) {
+    return transformService(originPath, destinationPath, functions, this);
   }
 
   /**